requirements.docker.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
APScheduler==3.11.0
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.3
django-apscheduler==0.7.0
et_xmlfile==2.0.0
frozenlist==1.7.0
geoip2==5.1.0
gunicorn==23.0.0
idna==3.10
maxminddb==2.7.0
multidict==6.6.3
numpy==2.2.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pillow==11.2.1
polib==1.2.0
propcache==0.3.2
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.4
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
stripe==12.2.0
tendo==0.3.0
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
xlsxwriter==3.2.5
yarl==1.20.1
selenium==4.25.0
webdriver-manager==4.0.2
